requirements.txt
streamlit==1.44.1
pillow==11.2.1
requests==2.32.3
opencv-python==4.11.0.86
